misc: uncrustify

+++ src/tools/timer.cpp
@@ -6,8 +6,8 @@
 CTimer::CTimer()
 {
 	timer_start = 0;
-	timer_end	= 0;
-	isStopped	= true;
+	timer_end   = 0;
+	isStopped   = true;
 }
 
 CTimer::~CTimer()
@@ -17,14 +17,14 @@ CTimer::~CTimer()
 void CTimer::start()
 {
 	timer_start = GetTickCount64();
-	timer_end	= 0;
-	isStopped	= false;
+	timer_end   = 0;
+	isStopped   = false;
 }
 
 void CTimer::stop()
 {
-	timer_end	= GetTickCount64();
-	isStopped	= true;
+	timer_end = GetTickCount64();
+	isStopped = true;
 }
 
 
@@ -36,7 +36,7 @@ unsigned CTimer::getElapsedMilliseconds()
 	return static_cast<unsigned>(timer_end - timer_start);
 }
 
-double CTimer::getElapsedSeconds()
+float CTimer::getElapsedSeconds()
 {
 	return getElapsedMilliseconds() / 1000.0;
 }
