+++ src/tools/timer.hpp
@@ -10,14 +10,14 @@ public:
 	void        stop();
 
 	unsigned    getElapsedMilliseconds();
-	double      getElapsedSeconds();
+	float       getElapsedSeconds();
 
 	bool operator()();
 
 
 private:
-	bool				isStopped;
-	unsigned long long	timer_start, timer_end;
+	bool               isStopped;
+	unsigned long long timer_start, timer_end;
 
 
 };
